Reorganize with package conventions
Need to confirm that this works properly with pipenv since I'm not so
familiar with the conventions
