+++ src/bayesian_inference/__init__.py
@@ -1,0 +1,2 @@
+
+__version__ = "0.1.0"
