+++ src/bayesian_inference/common_base.py
