+++ src/bayesian_inference/emulation.py
@@ -6,7 +6,7 @@ The main functionalities are:
  - fit_emulators() performs PCA, fits an emulator to each PC, and writes the emulator to file
  - predict() construct mean, std of emulator for a given set of parameter values
 
-A configuration class EmulationConfig provides simple access to emulation settings 
+A configuration class EmulationConfig provides simple access to emulation settings
 
 authors: J.Mulligan, R.Ehlers
 Based in part on JETSCAPE/STAT code.
@@ -21,8 +21,8 @@ import sklearn.preprocessing as sklearn_preprocessing
 import sklearn.decomposition as sklearn_decomposition
 import sklearn.gaussian_process as sklearn_gaussian_process
 
-import data_IO
-import common_base
+from bayesian_inference import data_IO
+from bayesian_inference import common_base
 
 ####################################################################################################################
 def fit_emulators(config):
@@ -45,7 +45,7 @@ def fit_emulators(config):
             return
 
     # Initialize predictions into a single 2D array: (design_point_index, observable_bins) i.e. (n_samples, n_features)
-    # A consistent order of observables is enforced internally in data_IO 
+    # A consistent order of observables is enforced internally in data_IO
     print(f'Doing PCA...')
     Y = data_IO.predictions_matrix_from_h5(config.output_dir, filename='observables.h5')
 
@@ -55,7 +55,7 @@ def fit_emulators(config):
     #
     # The input Y is a 2D array of format (n_samples, n_features).
     #
-    # The output of pca.fit_transform() is a 2D array of format (n_samples, n_components), 
+    # The output of pca.fit_transform() is a 2D array of format (n_samples, n_components),
     #   which is equivalent to:
     #     Y_pca = Y.dot(pca.components_.T), where:
     #       pca.components_ are the principal axes, sorted by decreasing explained variance -- shape (n_components, n_features)
@@ -88,7 +88,7 @@ def fit_emulators(config):
     kernel_matern = sklearn_gaussian_process.kernels.Matern(length_scale=length_scale,
                                                             length_scale_bounds=length_scale_bounds,
                                                             nu=1.5,
-                                                            )            
+                                                            )
     noise0 = 0.5**2
     noisemin = 0.01**2
     noisemax = 1**2
@@ -103,11 +103,11 @@ def fit_emulators(config):
     print()
     print(f'Fitting GPs...')
     print(f'  The design has {design.shape[1]} parameters')
-    emulators = [sklearn_gaussian_process.GaussianProcessRegressor(kernel=kernel, 
+    emulators = [sklearn_gaussian_process.GaussianProcessRegressor(kernel=kernel,
                                                              alpha=alpha,
                                                              n_restarts_optimizer=config.n_restarts,
                                                              copy_X_train=False).fit(design, y) for y in Y_pca_truncated.T]
-    
+
     # Print hyperparameters
     print()
     print('Kernel hyperparameters:')
@@ -137,9 +137,9 @@ def predict(parameters, results, config, validation_set=False):
     :return dict emulator_predicctions: dictionary of emulator predictions, with format emulator_predictions[observable_label]
     '''
 
-    # The emulators are stored as a list (one for each PC) 
+    # The emulators are stored as a list (one for each PC)
     emulators = results['emulators']
-            
+
     # Get predictions (in PC space) from each emulator and concatenate them into a numpy array with shape (n_design_points, n_PCs)
     emulator_mean = np.zeros((parameters.shape[0], config.n_pc))
     emulator_std = np.zeros((parameters.shape[0], config.n_pc))
@@ -150,7 +150,7 @@ def predict(parameters, results, config, validation_set=False):
         emulator_std[:,i] = y_std
 
     # Reconstruct the physical space from the PCs
-    pca = results['PCA']['pca'] 
+    pca = results['PCA']['pca']
     emulator_mean_reconstructed = emulator_mean.dot(pca.components_[:config.n_pc,:])
 
     # TODO: propagate and return emulator_std
